Clean up build script
Moves some constants - break classes i.a. - into a new source file
shared by both library code and the build script in order to reduce
duplication. This made it so that when building the pair table, the
break states can be referred to by their identifiers instead of doing
string lookups.