Restore compatibility with Rust 1.56
Rust 1.56 is the first version with support for the 2021 edition.
Maintaining support for this version gives people an easy target when
determining the oldest compiler they can use.

Textwrap tries to stay compatible with the same Rust version, and this
is where I first saw that the RSRV had increased to 1.59[1].

[1]: https://github.com/mgeisler/textwrap/issues/472